+++ spec/cable_server_helper.rb
@@ -1,1 +1,7 @@
+# frozen_string_literal: true
+
 CableServer.start
+
+RSpec.configure do |config|
+  config.after(:suite) { CableServer.stop }
+end